+++ backend/src/test/java/ca/metricalsky/winston/service/fetch/operation/FetchOperationHandlerTest.java
@@ -3,7 +3,7 @@ package ca.metricalsky.winston.service.fetch.operation;
 import ca.metricalsky.winston.entity.fetch.FetchOperationEntity;
 import ca.metricalsky.winston.exception.FetchOperationException;
 import ca.metricalsky.winston.service.fetch.FetchOperationService;
-import ca.metricalsky.winston.test.UnitTest;
+import ca.metricalsky.winston.test.annotations.UnitTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;