+++ frontend/src/components/channels/ChannelListToolbar.tsx
@@ -1,0 +1,30 @@
+import { ButtonGroup, ButtonToolbar, Col, Row } from "react-bootstrap";
+import { ShowArchivedChannelsButton } from "#/components/channels/ShowArchivedChannelsButton";
+import { pluralize } from "#/utils";
+import { PullChannelActions } from "./PullChannelActions";
+import type { ChannelListProps } from "#/types";
+
+type ChannelListToolbarProps = ChannelListProps & {
+  onPullChannel: () => void;
+  onRefreshChannels: () => void;
+};
+
+const ChannelListToolbar = (props: ChannelListToolbarProps) => {
+  return (
+    <Row className={"mb-2"}>
+      <Col className={"flex-center"}>{pluralize(props.channels.length, "channel")}</Col>
+      <Col xs={"auto"}>
+        <ButtonToolbar>
+          <ButtonGroup className={"me-2"}>
+            <ShowArchivedChannelsButton />
+          </ButtonGroup>
+          <ButtonGroup>
+            <PullChannelActions {...props} />
+          </ButtonGroup>
+        </ButtonToolbar>
+      </Col>
+    </Row>
+  );
+};
+
+export { ChannelListToolbar };