+++ frontend/src/routes/index.tsx
@@ -2,7 +2,7 @@ import { Route, Routes } from "react-router";
 import { routes } from "#/utils/links";
 import { AuthorListRoute } from "./authors";
 import { AuthorDetailsRoute } from "./authors/id";
-import { ChannelListRoute } from "./channels";
+import { ChannelListRoute } from "./channels/ChannelListRoute.tsx";
 import { ChannelDetailsRoute } from "./channels/id";
 import { VideoDetailsRoute } from "./videos/id";